riscv_plic

Struct Plic

pub struct Plic { /* private fields */ }

Platform-Level Interrupt Controller.

Implementations

impl Plic

pub const unsafe fn new(base: NonNull<PLICRegs>) -> Self

Create a new instance of the PLIC from the base address.

Safety

base must be a unique valid pointer to PLIC memory-mapped registers.

pub fn init_by_context(&mut self, ctx: usize)

Initialize the PLIC by context, setting the priority threshold to 0.

pub fn set_priority(&mut self, source: NonZeroU32, value: u32)

Sets priority for interrupt source to value.

Write 0 to priority value effectively disables this interrupt source, for the priority value 0 is reserved for “never interrupt” by the PLIC specification.

The lowest active priority is priority 1. The maximum priority depends on PLIC implementation and can be detected with Plic::probe_priority_bits.

See 4.

pub fn get_priority(&self, source: NonZeroU32) -> u32

Gets priority for interrupt source.

See 4.

pub fn probe_priority_bits(&mut self, source: NonZeroU32) -> u32

Probe maximum level of priority for interrupt source.

See 4.

pub fn is_pending(&self, source: NonZeroU32) -> bool

Check if interrupt source is pending.

See 5.

pub fn enable(&mut self, source: NonZeroU32, ctx: usize)

Enable interrupt source in context.

See 6.

pub fn disable(&mut self, source: NonZeroU32, ctx: usize)

Disable interrupt source in context.

See 6.

pub fn is_enabled(&self, source: NonZeroU32, ctx: usize) -> bool

Check if interrupt source is enabled in context.

See 6.

pub fn get_threshold(&self, ctx: usize) -> u32

Get interrupt threshold in context.

See 7.

pub fn set_threshold(&mut self, ctx: usize, value: u32)

Set interrupt threshold for context to value.

See 7.

pub fn probe_threshold_bits(&mut self, ctx: usize) -> u32

Probe maximum supported threshold value the context supports.

See 7.

pub fn claim(&mut self, ctx: usize) -> Option<NonZeroU32>

Claim an interrupt in context, returning its source.

It is always legal for a hart to perform a claim even if EIP is not set. A hart could set threshold to maximum to disable interrupt notification, but it does not mean interrupt source has stopped to send interrupt signals. In this case, hart would instead poll for active interrupt by periodically calling the claim function.

See 8.

pub fn complete(&mut self, ctx: usize, source: NonZeroU32)

Mark that interrupt identified by source is completed in context.

See 9.